Small changes after testing
The menu now allows you to make multiple commands without restarting the script;
A new option is added to explicitly quit the script as well.

Fixed a small error that made so the script would hang when setting the GPS antenna A offset and the GPS compass baseline.

Tested the script after changes and it works as expected